% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_plot.R
\name{cor_plot}
\alias{cor_plot}
\title{cor_plot: Frequency of occurrence of key covariates.}
\usage{
cor_plot(exp, outc, result)
}
\arguments{
\item{exp}{"exp" should be a character indicating the exposure variable, such as "A".}

\item{outc}{"outc" should be a character indicating the outcome variable, such as "B".}

\item{result}{"result" should be a data.frame obtained by the lm_cor/lm_auto_cor/logit_cor/logit_auto_cor functions.}
}
\description{
cor_plot: Frequency of occurrence of key covariates.
}
