% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_cor.R
\name{lm_cor}
\alias{lm_cor}
\title{lm_cor:Batch build linear regression models with specific combinations of covariates.}
\usage{
lm_cor(outc, exp, cor, data)
}
\arguments{
\item{outc}{outc should be a vector containing all outcome variables, such as c("A","B").}

\item{exp}{exp should be a vector containing all exposure variables, such as c("C","D").}

\item{cor}{cor should be a list containing a specific combination of covariates, such as list(c("E"),c("E","F")).}

\item{data}{data should be a data.frame containing all relevant variables.}
}
\description{
lm_cor:Batch build linear regression models with specific combinations of covariates.
}
