% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit_auto_cor.R
\name{logit_auto_cor}
\alias{logit_auto_cor}
\title{logit_auto_cor:Batch build logistic regression models with all possible combinations of covariates.}
\usage{
logit_auto_cor(outc, exp, cor, data, only_sig)
}
\arguments{
\item{outc}{outc should be a vector containing all outcome variables, such as c("A","B").}

\item{exp}{exp should be a vector containing all exposure variables, such as c("C","D").}

\item{cor}{cor should be a vector containing all covariates, such as c("E","F").}

\item{data}{data should be a data.frame containing all relevant variables.}

\item{only_sig}{only_sig should be TRUE or FALSE value.}
}
\description{
logit_auto_cor:Batch build logistic regression models with all possible combinations of covariates.
}
