% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit_cor.R
\name{logit_cor}
\alias{logit_cor}
\title{logit_cor:Batch build logistic regression models with specific combinations of covariates.}
\usage{
logit_cor(outc, exp, cor, data, only_sig)
}
\arguments{
\item{outc}{outc should be a vector containing all outcome variables, such as c("A","B").}

\item{exp}{exp should be a vector containing all exposure variables, such as c("C","D").}

\item{cor}{cor should be a list containing a specific combination of covariates, such as list(c("E"),c("E","F")).}

\item{data}{data should be a data.frame containing all relevant variables.}

\item{only_sig}{only_sig should be TRUE or FALSE value.}
}
\description{
logit_cor:Batch build logistic regression models with specific combinations of covariates.
}
