% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit_table.R
\name{logit_table}
\alias{logit_table}
\title{lm_table: Output logistic regression results as a table}
\usage{
logit_table(outc, exp, cor, data)
}
\arguments{
\item{outc}{"outc" should be a vector containing all outcome variables, such as c("A","B").}

\item{exp}{"exp" should be a vector containing all exposure variables, such as c("C","D").}

\item{cor}{"cor" should be a list containing a specific combination of covariates, such as list(c("E"),c("E","F")).}

\item{data}{"data" should be a data.frame containing all relevant variables.}
}
\description{
lm_table: Output logistic regression results as a table
}
