% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig_mod_num.R
\name{sig_mod_num}
\alias{sig_mod_num}
\title{sig_mod_num:Across all models, count the number of models in which the exposure variable was statistically significant.}
\usage{
sig_mod_num(exp, outc, result)
}
\arguments{
\item{exp}{exp should be a vector containing all exposure variables, such as c("C","D").}

\item{outc}{outc should be a vector containing all outcome variables, such as c("A","B").}

\item{result}{"result" should be a data.frame obtained by the lm_cor/lm_auto_cor/logit_cor/logit_auto_cor functions.}
}
\description{
sig_mod_num:Across all models, count the number of models in which the exposure variable was statistically significant.
}
